fix: make the oem field optional when retrieving the manager
DPUs seem to not return this field when the BMC fw is not upto date.
bump libredfish to 0.26.5

fix: capture the bios config job id when configuring the bios settings on Dells

fix: explicitly return an error if a BIOS config job is not generated in the machine_setup call. Rebooting the Dell after making a machine_setup call will be a no-op without generating the bios config job.